Use same pattern on iOS and android
