+++ lib/src/view/learn/learn_tab_screen.dart
@@ -11,7 +11,7 @@ import 'package:lichess_mobile/src/network/http.dart';
 import 'package:lichess_mobile/src/styles/styles.dart';
 import 'package:lichess_mobile/src/tab_scaffold.dart';
 import 'package:lichess_mobile/src/utils/l10n_context.dart';
-import 'package:lichess_mobile/src/view/account/account_drawer.dart';
+import 'package:lichess_mobile/src/view/account/account_menu.dart';
 import 'package:lichess_mobile/src/view/coordinate_training/coordinate_training_screen.dart';
 import 'package:lichess_mobile/src/view/study/study_list_screen.dart';
 import 'package:lichess_mobile/src/widgets/list.dart';
@@ -65,16 +65,13 @@ class LearnTabScreen extends ConsumerWidget {
       },
       child: PlatformScaffold(
         appBar: PlatformAppBar(
-          leading: Theme.of(context).platform == TargetPlatform.android
-              ? null
-              : const AccountDrawerIconButton(),
           title: Text(context.l10n.learnMenu),
-          centerTitle: Theme.of(context).platform != TargetPlatform.android,
-          actions: [
-            if (Theme.of(context).platform == TargetPlatform.android) const AndroidAccountButton(),
-          ],
+          centerTitle: false,
+          titleTextStyle: Theme.of(context).platform == TargetPlatform.iOS
+              ? Theme.of(context).textTheme.headlineMedium
+              : null,
+          actions: const [AccountMenuButton()],
         ),
-        drawer: Theme.of(context).platform == TargetPlatform.android ? null : const AccountDrawer(),
         body: const _Body(),
       ),
     );
