+++ lib/src/widgets/platform.dart
@@ -112,6 +112,7 @@ class PlatformAppBar extends StatelessWidget implements PreferredSizeWidget {
     this.leading,
     this.title,
     this.titleSpacing,
+    this.titleTextStyle,
     this.actions,
     this.bottom,
     this.centerTitle,
@@ -121,6 +122,7 @@ class PlatformAppBar extends StatelessWidget implements PreferredSizeWidget {
   final Widget? leading;
   final Widget? title;
   final double? titleSpacing;
+  final TextStyle? titleTextStyle;
   final List<Widget>? actions;
   final PreferredSizeWidget? bottom;
   final bool? centerTitle;
@@ -135,6 +137,7 @@ class PlatformAppBar extends StatelessWidget implements PreferredSizeWidget {
     final appBar = AppBar(
       titleSpacing: titleSpacing,
       title: title,
+      titleTextStyle: titleTextStyle,
       leading: leading,
       actions: actions,
       bottom: bottom,
